✨ feat(client): move userData to sessionStorage
refresh실행 시, 데이터가 날라가는 현상이 생겨 userinfo를 session으로 옮겼습니다
