+++ packages/client/package.json
@@ -27,6 +27,7 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.5",
+    "recoil-persist": "^4.2.0",
     "socket.io-client": "^4.5.2",
     "styled-components": "^5.3.5",
     "typescript": "^4.8.2",
