+++ packages/client/src/organisms/ChatSidebar/JoinedChatRoomList.tsx
@@ -8,8 +8,8 @@ import {
   ListUlLayout,
 } from '../../atoms/list/styles/ListStylesCSS';
 import UserProfileBox from '../../molecules/ProfileSection/UserProfileBox';
-import { userDataAtom } from '../../pages/PingpongRoutePage';
 import { chatRoomList, getJoinedChatList } from '../../recoil/chat.recoil';
+import { userDataAtom } from '../../recoil/user.recoil';
 import { chatNameSpace, enterChannel } from '../../socket/event';
 import useSocket from '../../socket/useSocket';
 import { UserProfileBoxDataType } from '../../types/Profile.type';
