+++ packages/client/src/template/ProfileSection/ProfileTemplate.tsx
@@ -13,10 +13,11 @@ import OtherInfo from '../../organisms/ProfileUserInfo/OtherInfo';
 import { BaseUserProfileData } from '../../types/Profile.type';
 import { useRecoilValue } from 'recoil';
 import { useParams } from 'react-router-dom';
-import { userDataAtom } from '../../pages/PingpongRoutePage';
+import { userDataAtom } from '../../recoil/user.recoil';
 
 function ProfileTemplate() {
   const user = useRecoilValue<BaseUserProfileData>(userDataAtom);
+  console.log('페이지마다 user정보확인', user);
   const { userId } = useParams();
   return (
     <ProfileLayout>
